Narrow weld original_to_effective from usize to u32
The original->effective index map is the larger half of the weld-fired
preprocess cost (a vec![0usize; n] zeroed and written per build). n is
bounded well below u32::MAX, so u32 halves its footprint (20MB->10MB at
2.5M) and the merge/compact bandwidth over it. This also makes the sphere
path consistent with the plane path, which already uses Vec<u32>.

Ripple is sphere-side only: MergeResult.original_to_effective, the
merge_result_from_pairs builder, compact_welded's parameter, and the two
consumers in compute.rs (error remap + cell remap, now `as usize` at the
indexing sites).

Preprocess on weld-firing sizes drops a further ~20% (2.5M: 85->67ms;
~65% below the pre-optimization 190ms). Bit-identity test and full suite
unchanged.

+++ src/cube_grid/weld.rs
@@ -158,7 +158,7 @@ impl CubeMapGrid {
     pub(crate) fn compact_welded(
         &mut self,
         kept: &[bool],
-        original_to_effective: &[usize],
+        original_to_effective: &[u32],
         n_eff: usize,
     ) {
         let num_cells = 6 * self.res * self.res;
@@ -180,7 +180,7 @@ impl CubeMapGrid {
                     dropped_slots.push(r as u32);
                     continue;
                 }
-                let eff = original_to_effective[orig] as u32;
+                let eff = original_to_effective[orig];
                 let (x, y, z) = (
                     self.cell_points_x[r],
                     self.cell_points_y[r],
